+++ src/main/java/com/finalproject/manitoone/domain/dto/NotificationResponse.java
@@ -1,7 +1,7 @@
-package com.finalproject.manitoone.notification.domain.dto;
+package com.finalproject.manitoone.domain.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.finalproject.manitoone.notification.constants.NotiType;
+import com.finalproject.manitoone.constants.NotiType;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
